refactor: separate into custom hook
